refactor(city-tabs): переносит логику компонента в отдельный хук
