+++ src/components/city-tabs/city-tabs.tsx
@@ -1,20 +1,8 @@
-import { MouseEvent } from 'react';
 import { CITIES } from '../../const';
-import { useAppDispatch } from '../../hooks/use-app-dispatch';
-import { useAppSelector } from '../../hooks/use-app-selector';
-import { fillPlacesList, setCity } from '../../store/action';
-import { City } from '../../types/city';
-import { SelectCity } from '../../store/selectors/city';
+import { useCity } from '../../hooks/use-city';
 
 function CityTabs(): JSX.Element {
-  const currentCity = useAppSelector(SelectCity);
-  const dispatch = useAppDispatch();
-
-  const cityChangeHandler = (evt: MouseEvent, city: City) => {
-    evt.preventDefault();
-    dispatch(setCity(city));
-    dispatch(fillPlacesList(city.name));
-  };
+  const {currentCity, cityChangeHandler} = useCity();
 
   return (
     <div className="tabs">
@@ -23,7 +11,7 @@ function CityTabs(): JSX.Element {
           {CITIES.map((city) => (
             <li className="locations__item" key={city.name}>
               <a className={`locations__item-link tabs__item ${(city.name === currentCity.name) ? 'tabs__item--active' : ''}`}
-                onClick={(evt) => cityChangeHandler(evt, city)}
+                onClick={cityChangeHandler(city)}
               >
                 <span>{city.name}</span>
               </a>
